Restructure imports to top-level entry points only
