+++ blocks/library/button/index.js
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-import IconButton from 'components/icon-button';
+import { IconButton } from 'components';
 
 /**
  * Internal dependencies
