+++ editor/sidebar/featured-image/index.js
@@ -8,10 +8,10 @@ import { connect } from 'react-redux';
  */
 import { Component } from 'element';
 import { __ } from 'i18n';
-import Button from 'components/button';
-import PanelBody from 'components/panel/body';
-import MediaUploadButton from 'blocks/media-upload-button';
-import Spinner from 'components/spinner';
+import { Button } from 'components';
+import { PanelBody } from 'components';
+import { MediaUploadButton } from 'blocks';
+import { Spinner } from 'components';
 
 /**
  * Internal dependencies
