+++ editor/sidebar/post-status/index.js
@@ -8,8 +8,8 @@ import { connect } from 'react-redux';
  */
 import { __ } from 'i18n';
 import { Component } from 'element';
-import PanelBody from 'components/panel/body';
-import FormToggle from 'components/form-toggle';
+import { PanelBody } from 'components';
+import { FormToggle } from 'components';
 
 /**
  * Internal Dependencies
