+++ editor/sidebar/post-trash/index.js
@@ -7,8 +7,8 @@ import { connect } from 'react-redux';
  * WordPress dependencies
  */
 import { __ } from 'i18n';
-import Button from 'components/button';
-import Dashicon from 'components/dashicon';
+import { Button } from 'components';
+import { Dashicon } from 'components';
 
 /**
  * Internal dependencies
